CRUD shapefile

+++ source/controllers/ShapefileController.js
@@ -33,14 +33,21 @@ class ShapefileController {
 
         const upload = multer(multerConfig).single('file');
         upload(req, res, async (fail) => {
+            
             if (fail instanceof multer.MulterError)
-                return res.status(400).send({ success: false, message: 'Houve um erro no processamento do arquivo!' });
+                return res.status(400).send({ success: false, message: 'Extensão de arquivo inválida!' });
+
+            if (!req.file)
+                return res.send({ success: false, message: 'É necessário submeter um arquivo!' });
+
 
             req.body = JSON.parse(req.body.data);
 
             const valid = ShapefileValidator.createValidate();
 
             const { error } = valid.validate(req.body);
+
+
 
             if (error) {
                 fs.unlinkSync(`${req.file.path}`);
@@ -52,15 +59,14 @@ class ShapefileController {
                 fs.unlinkSync(`${req.file.path}`);
                 return res.status(409).send({ success: false, message: 'Título já cadastrado!' });
             }
-            
+
             const newKey = req.file.key.split('.');
-            
+
             try {
-                console.log("entrou aqui")
                 await extract(req.file.path, { dir: `${req.file.destination}\\${newKey[0]}` });
 
                 fs.unlinkSync(`${req.file.path}`);
-                
+
             } catch (e) {
                 fs.unlinkSync(`${req.file.path}`);
 
@@ -75,7 +81,42 @@ class ShapefileController {
     }
 
     static async update(req, res) {
+        const valid = ShapefileValidator.updateValidate();
+        const { error } = valid.validate(req.body);
 
+        if (error)
+            return res.status(400).send({ success: false, message: error.details[0].message });
+
+
+        const { id, title, desc } = req.body;
+
+        const form = { id, title, desc };
+
+        const shp = await Shapefile.findOne(form.id);
+
+        if (shp.success) {
+            const toUpdate = {};
+
+            if (form.title && shp.shapefile.title != form.title) {
+                const existTitle = await Shapefile.findByTitle(form.title);
+                if (existTitle.success)
+                    return res.status(409).send({ success: false, message: 'Título já cadastrado!' });
+
+
+                toUpdate.title = form.title;
+            }
+
+            if (form.desc && shp.shapefile.desc != form.desc) toUpdate.desc = form.desc;
+
+            if (Object.keys(toUpdate).length) {
+                toUpdate.id = form.id;
+
+                const result = await Shapefile.update(toUpdate);
+                return result.success ? res.send(result) : res.status(400).send(result);
+            }
+            return res.send(shp);
+        }
+        return res.status(404).send({ success: false, message: 'Shapefile inexistente!' });
     }
 
     static async delete(req, res) {
@@ -85,12 +126,12 @@ class ShapefileController {
             return res.status(400).send({ success: false, message: 'Id inválido!' });
 
         // Verificar se o shapefile existe
-        shp = Shapefile.findOne(id);
+        const shp = await Shapefile.findOne(id);
         if (!shp.success)
             return res.status(404).send({ success: false, message: 'Shapefile inexistente!' });
 
 
-        const result = Shapefile.delete(id, shp.shapefile.uri);
+        const result = await Shapefile.delete(id, shp.shapefile.uri);
         result.success ? res.send(result) : res.status(400).send(result)
     }
 }
